Drop gtk-doc dependency
gtk-doc is not used anymore in mutter: gi-docgen is now used for
building the documentation.
Not replacing the dependency because API doc generation has been dropped
in commit 7b5a0bf5812dcb8b3d821d463b1f8ba353799c0d so gi-docgen is not
going to be used during the build anyway.

Signed-off-by: Sandro Bonazzola <sbonazzo@redhat.com>